==> jupyterNotebooks/00_darkskyWeatherAquisition.ipynb <==
import darksky as ds
import pandas as pd
import numpy as np

API_KEY = ''
BOSTON_LATITUDE = 42.355428
BOSTON_LONGITUDE = -71.069786
EXCLUDE=['flags', 'currently', 'minutely', 'alerts']

dates = pd.date_range(pd.Timestamp('2015-02-01'), pd.Timestamp('2016-11-30'), freq='D')

raw = []
for date in dates:
    date_iso = date.to_pydatetime().isoformat()
    response = ds.Forecast(key=API_KEY, latitude=BOSTON_LATITUDE, longitude=BOSTON_LONGITUDE, time=date_iso, units='si', exclude=EXCLUDE)
    raw[date_iso] = {
        'timezone': response['timezone'],
        'hourly': response['hourly']['data'],
        'daily': response['daily']['data']
    }

df_input = []
for day_index in output:
    for hour_data in output[day_index]['hourly']:
        hour_data['timestamp'] = pd.to_datetime(hour_data['time'], unit='s').tz_localize('UTC')
        df_input += [hour_data]

weather_df = pd.DataFrame(df_input)
weather_df = weather_df.set_index('timestamp')
weather_df = weather_df.tz_convert('EST')
weather_df = weather_df.drop(['time'], axis=1)

weather_df.to_csv('../data/raw/hourlyWeatherSummary_2015_2016.csv')
