==> jupyterNotebooks/01_tripDataPreparation.ipynb <==
import pandas as pd
import numpy as np
import math

trips = pd.read_csv('../data/raw/201501-hubway-tripdata.csv')
trips = trips.append(pd.read_csv('../data/raw/201502-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201503-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201504-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201505-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201506-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201507-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201508-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201509-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201510-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201511-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201512-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201601-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201602-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201603-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201604-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201605-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201606-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201607-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201608-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201609-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201610-hubway-tripdata.csv'))
trips = trips.append(pd.read_csv('../data/raw/201611-hubway-tripdata.csv'))
print(trips.shape)

trips.columns

#drop unwanted features and rename station features
trips = trips.drop(['tripduration', 'bikeid', 'usertype', 'birth year', 'gender'], axis=1)
trips = trips.rename(columns={'start station id': 'start_station_id', 
                              'start station name': 'start_station_name',
                              'start station latitude': 'start_station_latitude',
                              'start station longitude': 'start_station_longitude',
                              'end station id': 'end_station_id', 
                              'end station name': 'end_station_name',
                              'end station latitude': 'end_station_latitude',
                              'end station longitude': 'end_station_longitude'})

trips.dtypes

#coerce all end station ids to numbers. if not possible, then set NaN
trips['end_station_id'] = pd.to_numeric(trips['end_station_id'], errors='coerce')
print('Number of trips with unknown end station ID:', trips[pd.isnull(trips['end_station_id'])].shape[0])
trips = trips[pd.notnull(trips['end_station_id'])]
trips = trips.astype({'start_station_id': int, 'end_station_id': int})

#station ID 1 is Hubway's warehouse and can be removed
warehouse_trips = (trips['start_station_id'] == 1) | (trips['end_station_id'] == 1)
print('Number of trips with station ID = 1:', trips[warehouse_trips].shape[0])
trips = trips[warehouse_trips == False]

trips['end_station_latitude'] = pd.to_numeric(trips['end_station_latitude'], errors='coerce')
print('Number of trips with unknown end station latitude:', trips[pd.isnull(trips['end_station_latitude'])].shape[0])
trips['end_station_longitude'] = pd.to_numeric(trips['end_station_longitude'], errors='coerce')
print('Number of trips with unknown end station longitude:', trips[pd.isnull(trips['end_station_longitude'])].shape[0])

trips['starttime'] = trips['starttime'].apply(pd.Timestamp)
trips['stoptime'] = trips['stoptime'].apply(pd.Timestamp)
trips.dtypes

#remove trips from or to "virtual" locations outside Boston
BOSTON_LATITUDE = 42.355428
BOSTON_LONGITUDE = -71.069786
out_of_bounds = ((trips['start_station_latitude'] > (BOSTON_LATITUDE + 1)) | 
                 (trips['start_station_latitude'] < (BOSTON_LATITUDE - 1)) | 
                 (trips['end_station_latitude'] > (BOSTON_LATITUDE + 1)) | 
                 (trips['end_station_latitude'] < (BOSTON_LATITUDE - 1)) |
                 (trips['start_station_longitude'] > (BOSTON_LONGITUDE + 1)) | 
                 (trips['start_station_longitude'] < (BOSTON_LONGITUDE - 1)) | 
                 (trips['end_station_longitude'] > (BOSTON_LONGITUDE + 1)) | 
                 (trips['end_station_longitude'] < (BOSTON_LONGITUDE - 1)))
print("Number of trips with station locations outside of Boston: ", trips[out_of_bounds].shape[0])
trips = trips[out_of_bounds == False]

#remove trips that end in December 2016
print("Number of trips ending in December 2016: ", trips[trips['stoptime'] >= pd.Timestamp(2016, 12, 1, 0, 0, 0)].shape[0])
trips = trips[trips['stoptime'] < pd.Timestamp(2016, 12, 1, 0, 0, 0)]

trips.head()

print("Number of trips:", trips.shape[0])

trips = trips.set_index('starttime')
trips.to_csv('../data/raw/201501-201611-hubway-tripdata.csv')
