==> jupyterNotebooks/02_additionalDataPreparation.ipynb <==
import pandas as pd
import numpy as np
import math

#read weather data. its timestamp is already in date/hour format
weather = pd.read_csv('../data/raw/hourlyWeatherSummary_2015_2016.csv')
weather['date_hour'] = weather['timestamp'].apply(pd.Timestamp)
weather = weather.drop('timestamp', axis=1)

#add time features
weather['date'] = weather['date_hour'].dt.date
weather['hour'] = weather['date_hour'].dt.hour
weather['weekday'] = weather['date_hour'].dt.weekday
weather['month'] = weather['date_hour'].dt.month
weather['year'] = weather['date_hour'].dt.year
weather.head()

#read holiday data
holidays = pd.read_csv('../data/raw/publicHolidays_2015_2016.csv')
holidays['date'] = holidays['date'].apply(pd.Timestamp).dt.date
holidays.head()

#merge weather and holidays.
additional_features = weather.merge(holidays, how='left', on=('date'))

#add features to mark holidays and weekends
additional_features = additional_features.rename(columns={'description': 'holiday_description'})
additional_features['is_holiday'] = 1 * additional_features['holiday_description'].notnull()
additional_features['is_weekend'] = 1 * ((additional_features['weekday'] == 5) | (additional_features['weekday'] == 6))
additional_features['is_weekend_or_holiday'] = 1 * (additional_features['is_weekend'] | additional_features['is_holiday'])
additional_features = additional_features.set_index('date_hour')
additional_features.head()

additional_features.to_csv('../data/modelInput/additionalFeatures.csv')
