==> jupyterNotebooks/03_stationExtraction.ipynb <==
import pandas as pd
import numpy as np
import math
%matplotlib inline
import matplotlib.pyplot as plt
plt.rcParams['figure.figsize'] = [20, 10]
pd.options.mode.chained_assignment = None  # default='warn'

trips = pd.read_csv('../data/raw/201501-201611-hubway-tripdata.csv')
trips.head()

#convert timestamps
trips['starttime'] = trips['starttime'].apply(pd.Timestamp)
trips['stoptime'] = trips['stoptime'].apply(pd.Timestamp)
trips.dtypes

#split the columns into start and end data
#rename columns to a common format
start_stations = trips[['start_station_id', 'start_station_name', 'start_station_latitude', 'start_station_longitude', 'starttime']]
start_stations = start_stations.rename(columns={
    'start_station_id': 'station_id',
    'start_station_name': 'station_name', 
    'start_station_latitude': 'latitude', 
    'start_station_longitude': 'longitude',
    'starttime': 'timestamp'
})  

end_stations = trips[['end_station_id', 'end_station_name', 'end_station_latitude', 'end_station_longitude', 'stoptime']]
end_stations = end_stations.rename(columns={
    'end_station_id': 'station_id',
    'end_station_name': 'station_name', 
    'end_station_latitude': 'latitude', 
    'end_station_longitude': 'longitude',
    'stoptime': 'timestamp'
})  

#merge both sets into one
start_end_stations = start_stations.append(end_stations)

#save the station data. remove the timestamp and all duplicate IDs
stations = start_end_stations[['station_id', 'station_name', 'latitude', 'longitude']]
stations = stations.sort_values(by=['station_id'], ascending=True)
stations = stations.drop_duplicates(subset=['station_id'], keep='first')
stations = stations.set_index('station_id')
stations.head()

#reduce the full set of start and end data to station id and timestamp.
#sort by timestamp, with older timestamps on top
start_end_stations = start_end_stations[['station_id', 'timestamp']]
start_end_stations = start_end_stations.sort_values(by=['timestamp'], ascending=True)

#get the first entry for each station. it is the oldest departure or arrival timestamp
first_used = start_end_stations.drop_duplicates(subset=['station_id'], keep='first')

#convert the timestamp to date and drop the timestamp
first_used['first_used'] = first_used['timestamp'].dt.date
first_used = first_used.drop(['timestamp'], axis=1)
first_used = first_used.sort_values(by=['station_id'], ascending=True)
first_used = first_used.set_index('station_id')
first_used.head()

#same as above for the most recent arrival or departure
last_used = start_end_stations.drop_duplicates(subset=['station_id'], keep='last')

last_used['last_used'] = last_used.loc[:, ('timestamp')].dt.date
last_used = last_used.drop(['timestamp'], axis=1)
last_used = last_used.sort_values(by=['station_id'], ascending=True)
last_used = last_used.set_index('station_id')
last_used.head()

#join the datasets into one
stations = stations.join(first_used, how='outer')
stations = stations.join(last_used, how='outer')
stations.head()

print("Number of stations: ", stations.shape[0])

stations.to_csv('../data/modelInput/stations.csv')
