==> jupyterNotebooks/07_PCA.ipynb <==
import pandas as pd
import numpy as np
import math
import json
from sklearn.decomposition import PCA
from sklearn.externals import joblib
%matplotlib inline
import matplotlib.pyplot as plt
plt.rcParams['figure.figsize'] = [20, 10]

stations = pd.read_csv('../data/modelInput/stations_201505_201611.csv')
stations = stations.set_index('station_id')
stations['latitude_pca'] = 0
stations['longitude_pca'] = 0
stations.columns

#PCA for the coordinates of the reduced set of stations
pca_model = PCA()
stations[['latitude_pca', 'longitude_pca']] = pca_model.fit_transform(stations[['latitude', 'longitude']])
joblib.dump(pca_model, '../models/PCA.pkl') 

stations.to_csv('../data/modelInput/stations_201505_201611.csv')
